+++ src/components/DateField/index.tsx
@@ -1,1 +1,6 @@
-export { DateFieldOpenChangedEvent, DateFieldProps, CalendarViewOptions, DateField } from './DateFieldInt';
+export { DateField, DateFieldOpenChangedEvent, DateFieldProps } from './DateFieldInt.part';
+
+/**
+ * Deprecated
+ */
+export type CalendarViewOptions = 'days' | 'months' | 'years';